Put InSilicoRotater in own header file
